Fix validations failing when files are deleted from the validation directory
Files.walk() reads attributes from the files being walked. This requires the file to continue to exist between the directory scan and reading the attributes. If the file ceases to exist, an UncheckedIOException is thrown causing the validation to fail.

This commit removes reading file attributes except for testing if the path is a directory when discovering repositories. If the file stops existing Files.isDirectory() returns false and the validation continues.